APPLY_DIFF_simple
